Fixes rsa_pss_saltlen parameter for openssl/libressl
Older versions of openssl used this switch as follows:

For pss mode only this option specifies the salt length.  Two special values
are supported: -1 sets the salt length to the digest length.  When signing
-2 sets the salt length to the maximum permissible value.  When verifying -2
causes the salt length to be automatically determined based on the PSS block
structure.

Newer versions of openssl used this switch as follows:

For pss mode only this option specifies the salt length.  Three special
values are supported: digest sets the salt length to the digest length, max
sets the salt length to the maximum permissible value.  When verifying auto
causes the salt length to be automatically determined based on the PSS block
structure.